+++ src/core/logger/message.ts
